backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic-settings==2.7.0
youtube-transcript-api==0.6.3
httpx==0.28.1
google-generativeai==0.8.4
reportlab==4.2.5
python-multipart==0.0.18
PyJWT==2.10.1
pymongo==4.11.2
